fix: make numeric conversions explicit (bugprone-narrowing-conversions)
Resolve narrowing / implicit-widening / old-style-cast findings:

 - mavlink.cpp: cast scaled lat/lon to int32_t, velocities/altitudes to
   their MAVLink field types, command/interval to float, and the speed
   magnitude to uint16_t; make msecs_in_sec uint64_t so the back-off
   comparisons no longer widen an int multiplication; replace C-style
   socket casts with reinterpret_cast.
 - fss.cpp: cast the reported altitude to int16_t to match
   sendPosition()'s parameter type.
 - smm.cpp: cast the reported altitude to int32_t, the sweep width to
   int, and the waypoint count to int.

The MAVLink mission-item packer legitimately places scaled int32 lat/lon
next to a float altitude; bugprone-swapped-arguments cannot distinguish
them, so the two real coordinate-packing call sites are annotated with
NOLINTNEXTLINE. No behavioural change.

+++ src/smm/smm.cpp
@@ -113,8 +113,8 @@ SMM::reportPosition (PositionData t_pd)
         if (this->position_report_last_ts + 1000 <= curr_ts)
         {
             Point p = t_pd.getP ();
-            smm_asset_report_position (this->asset, p.getLatitude (), p.getLongitude (), t_pd.getAltitude (),
-                                       t_pd.getHeading () / 100, 3);
+            smm_asset_report_position (this->asset, p.getLatitude (), p.getLongitude (),
+                                       static_cast<int32_t> (t_pd.getAltitude ()), t_pd.getHeading () / 100, 3);
             this->position_report_last_ts = curr_ts;
         }
     }
@@ -226,13 +226,13 @@ SMMSearch::SMMSearch (smm_search t_search)
         this->addPoint (wp);
     }
     smm_waypoints_free (wps, wps_count);
-    this->altitude = smm_search_sweep_width (search);
+    this->altitude = static_cast<int> (smm_search_sweep_width (search));
 }
 
 auto
 SMMSearch::getPointsCount () -> int
 {
-    return this->points.size ();
+    return static_cast<int> (this->points.size ());
 }
 
 auto
